changed doubles to decimals

+++ Currencies/CourseDifferenceDto.cs
@@ -4,7 +4,7 @@ namespace XMLParse
 {
     public class CourseDifferenceDto
     {
-        public double Difference { get; set; }
+        public decimal Difference { get; set; }
         public DateTime Day { get; set; }
     }
 }
