+++ Currencies/CurrencyCourseInfoDto.cs
@@ -5,8 +5,8 @@ namespace XMLParse
     public class CurrencyCourseInfoDto
     {
         public string CurrencyCode { get; set; }
-        public double SellRate { get; set; }
-        public double BuyRate { get; set; }
+        public decimal SellRate { get; set; }
+        public decimal BuyRate { get; set; }
         public DateTime Day { get; set; }
     }
 }
